refactor: register logging callbacks in runner service
